Restructure parquet dir
